+++ src/main/kotlin/org/wfanet/measurement/duchy/deploy/postgres/writers/ClaimWork.kt
@@ -131,7 +131,12 @@ class ClaimWork<ProtocolT, StageT>(
     if (currentLockOwner.updateTime != unclaimedTask.updateTime) return false
 
     val writeTime = clock.instant()
-    acquireComputationLock(unclaimedTask.computationId, writeTime, ownerId, writeTime.plus(lockDuration))
+    acquireComputationLock(
+      unclaimedTask.computationId,
+      writeTime,
+      ownerId,
+      writeTime.plus(lockDuration)
+    )
     val stageLongValue =
       computationProtocolStagesEnumHelper
         .computationStageEnumToLongValues(unclaimedTask.computationStage)
