+++ src/main/kotlin/org/wfanet/measurement/duchy/service/internal/testing/ComputationsServiceTest.kt
@@ -414,7 +414,8 @@ abstract class ComputationsServiceTest<T : ComputationsCoroutineImplBase> {
       .isEqualTo(response.token)
   }
 
-//  @Test fun `updateComputationDetails updates RequisitionDetails`() = runBlocking { assert(false) }
+  //  @Test fun `updateComputationDetails updates RequisitionDetails`() = runBlocking {
+  // assert(false) }
 
   @Test
   fun `updateComputationDetails throws IllegalArgumentException when updating protocol`() =
